refactor : RaceRecords 객체 생성
