+++ src/main/java/racingcar/domain/RacingGame.java
@@ -2,26 +2,16 @@ package racingcar.domain;
 
 import racingcar.util.RandomNumberUtil;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
 public class RacingGame {
     private final int rounds;
     private final Cars cars;
-    private final List<RaceRecord> raceRecords;
+    private final RaceRecords raceRecords;
 
-    public RacingGame(Cars cars, int rounds) {
+    public RacingGame(Cars cars, RaceRecords raceRecords, int rounds) {
         this.cars = cars;
+        this.raceRecords = raceRecords;
         validateRounds(rounds);
         this.rounds = rounds;
-        this.raceRecords = new ArrayList<>();
-    }
-
-    public static RacingGame newRacingGame(Cars cars, int rounds) {
-        return new RacingGame(cars, rounds);
     }
 
     private void validateRounds(int rounds) {
@@ -29,11 +19,15 @@ public class RacingGame {
             throw new IllegalArgumentException("시도 횟수는 1회 이상이어야 합니다 :) ");
     }
 
+    public static RacingGame newRacingGame(Cars cars, RaceRecords raceRecords, int rounds) {
+        return new RacingGame(cars, raceRecords, rounds);
+    }
+
     public int getRounds() {
         return rounds;
     }
 
-    public List<RaceRecord> getRaceRecords() {
+    public RaceRecords getRaceRecords() {
         return raceRecords;
     }
 
@@ -45,36 +39,6 @@ public class RacingGame {
 
     private void playSingleRound(int round) {
         cars.getCars().forEach(c -> c.tryToMove(RandomNumberUtil.getRandomNumber()));
-        raceRecords.add(new RaceRecord(round, saveRecords()));
+        raceRecords.getRaceRecords().add(new RaceRecord(round, raceRecords.saveRecords(cars)));
     }
-
-    private Map<String, Integer> saveRecords() {
-        Map<String, Integer> map = new HashMap<>();
-        for (Car car : cars.getCars()) {
-            map.put(car.getName(), car.getLocation());
-        }
-        return map;
-    }
-
-    public String getWinner() {
-        RaceRecord lastRaceRecord = getLastRaceRecord();
-        Map<String, Integer> recordsWithCarName = lastRaceRecord.getRecordsWithCarName();
-        Integer max = Collections.max(recordsWithCarName.values());
-        return getWinnerNames(recordsWithCarName, max);
-    }
-
-    private RaceRecord getLastRaceRecord() {
-        return getRaceRecords().get(getRaceRecords().size() - 1);
-    }
-
-    private String getWinnerNames(Map<String, Integer> map, Integer max) {
-        StringBuilder winners = new StringBuilder();
-        for (String name : map.keySet()) {
-            if (map.get(name).equals(max)) {
-                winners.append(name).append(", ");
-            }
-        }
-        return winners.toString();
-    }
-
 }
