+++ src/main/java/racingcar/view/ResultView.java
@@ -8,12 +8,12 @@ import java.util.List;
 
 public class ResultView {
     public void print(RacingGame racingGame) {
-        List<RaceRecord> raceRecords = racingGame.getRaceRecords();
+        List<RaceRecord> raceRecords = racingGame.getRaceRecords().getRaceRecords();
         for (RaceRecord raceRecord : raceRecords) {
             System.out.println("ROUND " + raceRecord.getRound());
             raceRecord.getRecordsWithCarName()
                     .forEach((key, value) -> System.out.printf("%s : %s \n", key, StateUtil.makeStateLine(value)));
         }
-        System.out.println(racingGame.getWinner() + "가 우승했습니다.");
+        System.out.println(racingGame.getRaceRecords().getWinner() + "가 우승했습니다.");
     }
 }
